# Fix and normalize Vector4 UnPremultiply