--- a/src/ImageSharp/PixelFormats/PixelBlenders/PorterDuffFunctions.Generated.tt
+++ b/src/ImageSharp/PixelFormats/PixelBlenders/PorterDuffFunctions.Generated.tt
@@ -33,7 +33,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>Src(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return source;
     }
@@ -59,7 +59,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>SrcAtop(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Atop(backdrop, source, <#=blender#>(backdrop, source));
     }
@@ -89,7 +89,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>SrcOver(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Over(backdrop, source, <#=blender#>(backdrop, source));
     }
@@ -119,7 +119,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>SrcIn(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return In(backdrop, source);
     }
@@ -145,7 +145,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>SrcOut(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Out(backdrop, source);
     }
@@ -197,7 +197,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>DestAtop(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Atop(source, backdrop, <#=blender#>(source, backdrop));
     }
@@ -227,7 +227,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>DestOver(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Over(source, backdrop, <#=blender#>(source, backdrop));
     }
@@ -257,7 +257,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>DestIn(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return In(source, backdrop);
     }
@@ -283,7 +283,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>DestOut(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Out(source, backdrop);
     }
@@ -309,7 +309,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>Xor(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Xor(backdrop, source);
     }
@@ -335,7 +335,7 @@ internal static partial class PorterDuffFunctions
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static Vector4 <#=blender#>Clear(Vector4 backdrop, Vector4 source, float opacity)
     {
-        source = WithW(source, source * opacity);
+        source = Numerics.WithW(source, source * opacity);
 
         return Clear(backdrop, source);
     }