--- a/tests/ImageSharp.Benchmarks/Color/Bulk/UnPremultiplyVector4.cs
+++ b/tests/ImageSharp.Benchmarks/Color/Bulk/UnPremultiplyVector4.cs
@@ -8,7 +8,6 @@ using BenchmarkDotNet.Attributes;
 
 namespace SixLabors.ImageSharp.Benchmarks.ColorSpaces.Bulk;
 
-[Config(typeof(Config.ShortCore31))]
 public class UnPremultiplyVector4
 {
     private static readonly Vector4[] Vectors = CreateVectors();
@@ -26,10 +25,7 @@ public class UnPremultiplyVector4
     }
 
     [Benchmark]
-    public void UnPremultiply()
-    {
-        Numerics.UnPremultiply(Vectors);
-    }
+    public void UnPremultiply() => Numerics.UnPremultiply(Vectors);
 
     [MethodImpl(InliningOptions.ShortMethod)]
     private static void UnPremultiply(ref Vector4 source)
@@ -41,13 +37,13 @@ public class UnPremultiplyVector4
 
     private static Vector4[] CreateVectors()
     {
-        var rnd = new Random(42);
+        Random rnd = new(42);
         return GenerateRandomVectorArray(rnd, 2048, 0, 1);
     }
 
     private static Vector4[] GenerateRandomVectorArray(Random rnd, int length, float minVal, float maxVal)
     {
-        var values = new Vector4[length];
+        Vector4[] values = new Vector4[length];
 
         for (int i = 0; i < length; i++)
         {