… src/ImageSharp/PixelFormats/PixelBlenders/PorterDuffFunctions.Generated.cs
… tests/ImageSharp.Tests/Formats/Tiff/TiffDecoderTests.cs
… tests/Images/External/ReferenceOutput/PorterDuffCompositorTests/PorterDuffOutputIsCorrect_Rgba32_pd-dest_DestOut.png
… tests/Images/External/ReferenceOutput/PorterDuffCompositorTests/PorterDuffOutputIsCorrect_Rgba32_pd-dest_In.png
… tests/Images/External/ReferenceOutput/PorterDuffCompositorTests/PorterDuffOutputIsCorrect_Rgba32_pd-dest_Out.png