### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description
Our existing `UnPremultiply` methods suffered from an issue in that they allowed dividing values by zero. This was masked by our tests since the transform to byte would lead to `zero` in most cases but could cause issues when chaining certain operations requiring premultiplication using single precision pixel formats since interim values for the `XYZ` components would be `∞`.

The Porter-Duff operators contained a bad fix for this by dividing by the max of the `W` component vs an epsilon of `0.001F`. This would lead to out-of-range values if the X, Y, Z components had a non-zero value. This was generally masked by clamping but again could cause issues for single precision pixel formats.

The fix is simple. Compare the alpha component to `zero` and return the source value if so. We return the source rather than an explicit zeroed value for two reasons:

1. For all our operation this would be equivalent to `Vector4.Zero`
2. There are Tiff reference files in our tests suite that are encoded with associated alpha values of `zero` but have `non-zero` RGB component values. This allows us to maintain the integrity of those values. (_Note the reference decoder makes the same mistake we did before this fix_).

The fix does not affect the performance of `UnPremultiply`. Migrating the Porter-Duff operators to use the fixed method however yields an additional 1.55x performance boost over the work done in #2359 

This brings an overall performance boost of 14.4x over our v2 version of the operators.

```bash
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.22621
11th Gen Intel Core i7-11370H 3.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK=7.0.200
  [Host]     : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT
  DefaultJob : .NET 6.0.14 (6.0.1423.7309), X64 RyuJIT
```

Porter-Duff NormalSrcOver

Main
| Method |     Mean |   Error |  StdDev | Ratio |
|------- |---------:|--------:|--------:|------:|
| Scalar | 416.1 ns | 0.86 ns | 0.80 ns |  1.00 |
|    Avx | 235.3 ns | 0.62 ns | 0.55 ns |  0.57 |

PR

| Method |     Mean |   Error |  StdDev | Ratio |
|------- |---------:|--------:|--------:|------:|
| Scalar | 354.6 ns | 0.56 ns | 0.50 ns |  1.00 |
|    Avx | 151.4 ns | 0.43 ns | 0.40 ns |  0.43 |

PremultiplyVector4

Main

|              Method |     Mean |    Error |   StdDev | Ratio |
|-------------------- |---------:|---------:|---------:|------:|
| PremultiplyBaseline | 31.16 us | 0.061 us | 0.054 us |  1.00 |
|         Premultiply | 22.95 us | 0.050 us | 0.044 us |  0.74 |

PR

|              Method |     Mean |    Error |   StdDev | Ratio |
|-------------------- |---------:|---------:|---------:|------:|
| PremultiplyBaseline | 31.17 us | 0.074 us | 0.066 us |  1.00 |
|         Premultiply | 22.94 us | 0.042 us | 0.039 us |  0.74 |


UnPremultiplyVector4

Main

|                Method |     Mean |     Error |    StdDev | Ratio |
|---------------------- |---------:|----------:|----------:|------:|
| UnPremultiplyBaseline | 1.539 us | 0.0023 us | 0.0020 us |  1.00 |
|         UnPremultiply | 1.265 us | 0.0015 us | 0.0013 us |  0.82 |

PR

|                Method |     Mean |     Error |    StdDev | Ratio |
|---------------------- |---------:|----------:|----------:|------:|
| UnPremultiplyBaseline | 1.538 us | 0.0031 us | 0.0029 us |  1.00 |
|         UnPremultiply | 1.266 us | 0.0013 us | 0.0012 us |  0.82 |

<!-- Thanks for contributing to ImageSharp! -->

One thing I do not understand though is why is there such a difference between `Premultiply` vs `UnPremultiply`

The baseline implementations in the benchmark use virtually identical code and generate [virtually identical ASM](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgDkBXfGKASzFwBuGjWIBmJqQYBhBgG8aDJQwAOfAG7YMMJuSRMUDAAqx8HADYZeK8wE8AFLABmDAGowwGaIdwQOUMBgASkVlBWplSIYncwgtBgB3BgBeBl9/QJYAdWEIqKV0gJ0AKlSE3PyCvyLslMSK5QBfUKUW1Q0tHWZ9YkMAVQA7ExgzS2s7RxgXd09vNOrAkLylcMro2Pik1MLMnLbInZ0AejKG/MPa07bm6kagA).

@tannergooding would you have any insight? 

```c#
private static void Premultiply(ref Vector4 source)
{
    float w = source.W;
    source *= w;
    source.W = w;
}

private static void UnPremultiply(ref Vector4 source)
{
    float w = source.W;
    source /= w;
    source.W = w;
}
```

```asm
C.Premultiply(System.Numerics.Vector4 ByRef)
    L0000: vzeroupper
    L0003: vmovss xmm0, [rcx+0xc]
    L0008: vmovupd xmm1, [rcx]
    L000c: vbroadcastss xmm2, xmm0
    L0011: vmulps xmm1, xmm1, xmm2
    L0015: vmovupd [rcx], xmm1
    L0019: vmovss [rcx+0xc], xmm0
    L001e: ret

C.UnPremultiply(System.Numerics.Vector4 ByRef)
    L0000: vzeroupper
    L0003: vmovss xmm0, [rcx+0xc]
    L0008: vmovupd xmm1, [rcx]
    L000c: vbroadcastss xmm2, xmm0
    L0011: vdivps xmm1, xmm1, xmm2
    L0015: vmovupd [rcx], xmm1
    L0019: vmovss [rcx+0xc], xmm0
    L001e: ret
```