@tannergooding 2023-02-24
> One thing I do not understand though is why is there such a difference between Premultiply vs UnPremultiply

Floating-point division is quite a bit slower than multiplication (although not quite as bad as integral division).

`divps` takes around `11-12 cycles` to execute (its latency) and has about `3 cycles` before another non-dependent instruction can execute on the same port (reciprocal throughput).

`mulps` on the other hand takes `3-4 cycles` to execute and has about `0.5 cycles` before another non-dependent instruction can execute on the same port.

----------

This is notably another case where you'd probably see some benefit in using the `PermuteW` and `WithW` helpers.

What you have now is roughly:
```asm
C.Premultiply(System.Numerics.Vector4 ByRef)
    L0000: vzeroupper                   ;
    L0003: vmovss xmm0, [rcx+0xc]       ; 4-7 cycles;  port 2 or 3
    L0008: vmovupd xmm1, [rcx]          ; 4-7 cycles;  port 2 or 3
    L000c: vbroadcastss xmm2, xmm0      ; 1 cycle;     port 5
    L0011: vmulps xmm1, xmm1, xmm2      ; 4 cycles;    port 0 or 1
    L0015: vmovupd [rcx], xmm1          ; 4-10 cycles; port 2 or 3 or 7 + port 4
    L0019: vmovss [rcx+0xc], xmm0       ; 4-10 cycles; port 2 or 3 or 7 + port 4
    L001e: ret                          ;
```

Given dependencies, parallel dispatch, and other bits this will take around `17-39` cycles.

However, if you change the implementation to something like:
```csharp
private static void Premultiply1(ref Vector4 source)
{
    var vsource = source.AsVector128();

    Vector4 w = PermuteW(vsource);
    vsource *= w;
    source = WithW(vsource, w);
}
```

We instead generate:
```asm
    L0000: vzeroupper                           ;
    L0003: vmovupd xmm0, [rcx]                  ; 4-7 cycles;  port 2 or 3
    L0007: vpermilps xmm1, xmm0, 0xff           ; 1 cycle;     port 5
    L000d: vmulps xmm0, xmm0, xmm1              ; 4 cycles;    port 0 or 1
    L0011: vinsertps xmm0, xmm0, xmm1, 0xf0     ; 1 cycle;     port 5
    L0017: vmovupd [rcx], xmm0                  ; 4-10 cycles; port 2 or 3 or 7 + port 4
    L001b: ret                                  ;
```

Which is around `14-23` cycles instead, which avoids the extra memory accesses and uses less overall ports.