@JimBobSquarePants 2023-02-24
> > One thing I do not understand though is why is there such a difference between Premultiply vs UnPremultiply
> 
> Floating-point division is quite a bit slower than multiplication (although not quite as bad as integral division).
> 
> `divps` takes around `11-12 cycles` to execute (its latency) and has about `3 cycles` before another non-dependent instruction can execute on the same port (reciprocal throughput).
> 
> `mulps` on the other hand takes `3-4 cycles` to execute and has about `0.5 cycles` before another non-dependent instruction can execute on the same port.
> 

Thanks for all the info there, much appreciated.

What I'm seeing though is that the division is faster. A lot faster. 1.538 us vs 31.17 us. It makes no sense to me.