@tannergooding 2023-02-24
> What I'm seeing though is that the division is faster. A lot faster. 1.538 us vs 31.17 us. It makes no sense to me.

Ah yes. I misread the numbers initially.

This isn't clear to me either and I'd probably end up doing a profile using `Intel VTune` or `AMD uProf` so I could get instruction level timing information at this point.

My gut instinct would be a difference in input data or in how BDN ir measuring the two profiles.